+++ Examples/Example02/Example02-Customized.ps1
@@ -1,2 +1,0 @@
-New-BurntToastNotification -AppLogo C:\smile.jpg -Text "Don't forget to smile!",
-                                                       'Your script ran successfully, celebrate!'