+++ CHANGES.md
@@ -12,6 +12,10 @@
 
 - Ensure all invalid characters are removed from potential file system paths when caching images (thanks @markekraus)
 
+### Libraries
+
+- Microsoft.Windows.SDK.NET.Ref libraries bumped to 10.0.22621.26
+
 ## [v1.0.0-Preview1](https://github.com/Windos/BurntToast/releases/download/v1.0.0-Preview1/BurntToast.zip)
 
 ### Breaking Changes